Map click listener added

+++ pages/1_Map.py
@@ -1,19 +1,23 @@
 import os
+from typing import List
+
 import pandas as pd
+import plotly.express as px
 import streamlit as st
+from plotly.graph_objects import Figure
+from sklearn.manifold import TSNE
+from streamlit_plotly_events import plotly_events
+
 from widgets.console import show_console
-from widgets.session import app_config
+from widgets.download import download_dataframe_button
 from widgets.index import (
     get_last_index_update,
     load_codelist_reverse_index,
     load_table_of_contents,
 )
-from widgets.download import download_dataframe_button
-from sklearn.manifold import TSNE
-import plotly.express as px
+from widgets.session import app_config
 
 
-@st.experimental_memo
 def build_labeled_toc() -> pd.DataFrame:
     # Datasets starts with a code that identify its theme like:
     # aact_ali01 (Agricultural labour input statistics: absolute...) ->	aact (Economic accounts for agriculture)
@@ -30,7 +34,6 @@ def build_labeled_toc() -> pd.DataFrame:
     return ab
 
 
-@st.experimental_memo
 def build_adjacency_matrix() -> pd.DataFrame:
     # Reverse variable index (var -> List(dataset)) can build an adjacency matrix to
     # describe a dataset by the variable used by it.
@@ -45,18 +48,19 @@ def build_adjacency_matrix() -> pd.DataFrame:
 
 
 @st.experimental_memo
-def cluster_datasets(adjacency: pd.DataFrame, toc: pd.DataFrame) -> pd.DataFrame:
+def cluster_datasets() -> pd.DataFrame:
+    toc = build_labeled_toc()
+    adj = build_adjacency_matrix()
     # Project datasets into 2D space
     tsne = TSNE(n_components=2, learning_rate="auto", metric="cosine", init="pca")
-    xy = pd.DataFrame(
-        tsne.fit_transform(adjacency), index=adjacency.index, columns=["1st", "2nd"]
-    )
+    xy = pd.DataFrame(tsne.fit_transform(adj), index=adj.index, columns=["1st", "2nd"])
     xy.index.name = "code"
     # Join datasets with label
     return xy.join(toc).reset_index()
 
 
-def show_clustering(data: pd.DataFrame, margin: int = 5):
+@st.experimental_memo
+def plot_clustering(data: pd.DataFrame, margin: int = 5) -> Figure:
     fig = px.scatter(
         data.rename(columns={"title_theme": "Themes"}),
         x="1st",
@@ -73,35 +77,54 @@ def show_clustering(data: pd.DataFrame, margin: int = 5):
         ),
         height=1300,
     )
-    fig.update_layout(legend=dict(orientation="h"))
-    st.plotly_chart(figure_or_data=fig, use_container_width=True)
+    fig = fig.update_layout(legend=dict(orientation="h"))
+    # Keep zoom at click: https://discuss.streamlit.io/t/cant-enter-values-without-updating-a-plotly-figure/28066
+    fig = fig.update_layout({"uirevision": "foo"}, overwrite=True)
+    return fig
+
+
+@st.experimental_memo
+def coordinates_as_index(data: pd.DataFrame) -> pd.DataFrame:
+    return data.set_index(["1st", "2nd"])
 
 
 if __name__ == "__main__":
     app_config("Map")
 
     st.header("Datasets map")
-    if os.environ["ENV"] == "streamlit":
-        dataset2d_path = "cache/clustermap.csv.gz"
-        if not os.path.exists(dataset2d_path):
-            st.error(
-                "Datasets clustering is too expensive for Streamlit Cloud limited resources. You can compute this offline, cloning the repo."
-            )
-            buffer = st.file_uploader("Load clustering offline results", "gz")
-            if buffer:
-                with open(dataset2d_path, "wb") as f:
-                    f.write(buffer.getbuffer())
-                    st.experimental_rerun()
-        else:
-            show_clustering(pd.read_csv(dataset2d_path))
+
+    dataset2d_path = "cache/clustermap.csv.gz"
+    if os.environ["ENV"] == "streamlit" and not os.path.exists(dataset2d_path):
+        st.error(
+            "Datasets clustering is too expensive for Streamlit Cloud limited resources. You can compute this offline, cloning the repo."
+        )
+        buffer = st.file_uploader("Load clustering offline results", "gz")
+        if buffer:
+            with open(dataset2d_path, "wb") as f:
+                f.write(buffer.getbuffer())
+                st.experimental_rerun()
     else:
         if not get_last_index_update():
             st.warning("Create an index first!")
         else:
-            labeled_toc = build_labeled_toc()
-            adj = build_adjacency_matrix()
-            datasets2d = cluster_datasets(adj, labeled_toc)
-            show_clustering(datasets2d)
+            datasets2d = (
+                pd.read_csv(dataset2d_path)
+                if os.environ["ENV"] == "streamlit"
+                else cluster_datasets()
+            )
+            selection = plotly_events(
+                plot_clustering(datasets2d),
+                click_event=True,
+                select_event=True,
+                # use_container_width=True,  # Not supported by plotly_events
+                override_height=600,
+                override_width="100%",
+            )
+            if selection:
+                selection = [(s["x"], s["y"]) for s in selection]
+                st.sidebar.subheader("Selection")
+                st.sidebar.dataframe(coordinates_as_index(datasets2d).loc[selection])
+
             download_dataframe_button(datasets2d, filename_prefix="clustermap")
 
     show_console()  # For debugging
